requirements.txt
Flask==3.1.1
python-dotenv==1.0.1
google-generativeai==0.8.5
gunicorn==22.0.0
